refactor(racing-preparation): 경기 시도 횟수 BigInteger에서 int로 변경

+++ src/main/java/racingcar/RacingSimulatorController.java
@@ -1,6 +1,5 @@
 package racingcar;
 
-import java.math.BigInteger;
 import java.util.List;
 
 public class RacingSimulatorController {
@@ -27,7 +26,7 @@ public class RacingSimulatorController {
         String racingAttempt = consoleReader.readLine();
 
         List<RacingCar> racingCars = RacingCarFactory.racingCars(carNames);
-        BigInteger racingAttemptNumber = AttemptParser.parse(racingAttempt);
+        int racingAttemptNumber = AttemptParser.parse(racingAttempt);
     }
 
     public void stop() {
